requirements.txt
PyQt6
google-generativeai
pdfplumber
python-docx
requests
SpeechRecognition
pyaudio